Add files via upload
Gross Domestic Product(GDP) is a measure to gauge the health of an economy and represents the total value of goods and services produced in an economy. GDP growth rate is the major indicator of a country’s economic performance. Per capita GDP is a global indicator of a country’s economy that economists use in combination with GDP to assess a country’s wealth based on its economic growth. The formula of GDP per capita is:

GDP per capita = Gross Domestic Product (GDP) / Population